Adopt to relocated packages in Spring Data Cassandra.
