+++ spring-boot-autoconfigure/src/main/java/org/springframework/boot/autoconfigure/data/cassandra/CassandraReactiveRepositoriesAutoConfiguration.java
@@ -21,18 +21,19 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.cassandra.core.session.ReactiveSession;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.data.cassandra.repository.ReactiveCassandraRepository;
 import org.springframework.data.cassandra.repository.config.EnableReactiveCassandraRepositories;
 import org.springframework.data.cassandra.repository.support.ReactiveCassandraRepositoryFactoryBean;
+import org.springframework.data.cql.core.session.ReactiveSession;
 
 /**
  * {@link EnableAutoConfiguration Auto-configuration} for Spring Data's Cassandra Reactive
  * Repositories.
  *
  * @author Eddú Meléndez
+ * @author Mark Paluch
  * @since 2.0.0
  * @see EnableReactiveCassandraRepositories
  */
